+++ tests/data_model_test.py
@@ -1,3 +1,5 @@
+from pathlib import Path
+
 import pytest
 
 from beaker.data_model import *
@@ -27,21 +29,26 @@ def test_data_source_validation():
     assert DataSource(host_path="bar").host_path == "bar"
 
 
-def test_experiment_spec_from_and_to_json(beaker_cluster_name):
+def test_experiment_spec_from_and_to_json_and_file(beaker_cloud_cluster_name: str, tmp_path: Path):
     json_spec = {
         "version": "v2",
         "tasks": [
             {
                 "name": "main",
                 "image": {"docker": "hello-world"},
-                "context": {"cluster": beaker_cluster_name},
+                "context": {"cluster": beaker_cloud_cluster_name},
                 "result": {"path": "/unused"},
                 "resources": {"memory": "512m", "sharedMemory": "512m"},
             },
         ],
     }
+
     spec = ExperimentSpec.from_json(json_spec)
     assert spec.to_json() == json_spec
+
+    spec_path = tmp_path / "spec.yml"
+    spec.to_file(spec_path)
+    assert ExperimentSpec.from_file(spec_path) == spec
 
 
 def test_experiment_spec_validation():